# contracts/openapi/domain-services.v1.yaml
openapi: 3.1.0
info:
  title: SurakshaSetu domain services
  version: 1.0.0
  description: |
    The tier contract between the orchestrator (Python) and the domain services (Java). It is the
    single source of truth: Java interfaces and models and Python models are generated from it.
    Model shapes follow the TDD (§3.5–§3.8) where it defines them.

    ## Conventions
    - **Money** is a decimal string of rupees (`^\d+(\.\d{1,2})?$`), never a JSON number.
    - **Dates** are ISO 8601 (`date` or `date-time`); date-times are UTC.
    - **Errors** are `application/problem+json` with `type`, `title`, `status`, `detail` and a
      machine-readable `code`.
    - **Auth**: a bearer token (a shared dev token locally; mTLS in production).
    - **Decisions** (`POST` eligibility, suitability, ranking and quotes): the request carries
      `pins.rules` and the service evaluates exactly that version, or answers 409
      `RULES_VERSION_UNKNOWN`. The response carries `decision_id`, the version it used
      (`rules_version`, `ranker_version` or `rating_version`), `inputs_sha256` and `reason_codes`.
    - Commission, margin, incentive, campaign and payout fields never appear in this contract.

    ## Problem codes
    | Status | code | When |
    | --- | --- | --- |
    | 404 | `NOT_FOUND` | Unknown UIN, consent id, notice, disclosure, pincode or occupation. |
    | 409 | `RULES_VERSION_UNKNOWN` | `pins.rules` names a version the service has not loaded. |
...

    ## Hashing
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/catalog/products:
    get:
      tags: [catalog]
      operationId: listProducts
      parameters:
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/ProductStatus'}
        - name: category
          in: query
          schema: {$ref: '#/components/schemas/ProductType'}
        - name: launch_enabled
          in: query
          schema: {type: boolean}
        - $ref: '#/components/parameters/AsOf'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Product'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/catalog/products/{uin}:
    get:
      tags: [catalog]
      operationId: getProduct
      summary: A product with its riders and documents. A withdrawn product is served only for an as_of before its effective_to.
      parameters:
        - $ref: '#/components/parameters/Uin'
        - $ref: '#/components/parameters/AsOf'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Product'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/catalog/riders/{uin}:
    get:
      tags: [catalog]
      operationId: getRider
      parameters:
        - $ref: '#/components/parameters/Uin'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Rider'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/catalog/products/{uin}/kill-switch:
    post:
      tags: [catalog]
      operationId: setProductKillSwitch
      summary: Withdraw a product from sale immediately. Internal scope.
      parameters:
        - $ref: '#/components/parameters/Uin'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/KillSwitch'}
      responses:
        '200':
          description: Kill switch set; the product's new status.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KillSwitchResult'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    AsOf:
      name: as_of
      in: query
      description: Evaluate as of this instant. Defaults to now.
      schema: {type: string, format: date-time}
    Uin:
      name: uin
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uin'}
...
  responses:
    Problem:
      description: Error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    Problem:
      type: object
      required: [title, status, code]
      properties:
        type: {type: string, default: 'about:blank', description: 'RFC 9457: absent means about:blank.'}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance:
          type: string
          description: 'RFC 9457: the request path. It can carry ids or a pincode, so never log it.'
        code:
          type: string
          description: Machine-readable error code; the table in the API description lists them.
    Money:
      type: string
      description: Rupees as a decimal string.
      pattern: &money '^\d+(\.\d{1,2})?$'
...
    Sha256Hex:
      type: string
      pattern: '^[0-9a-f]{64}$'
    Uin:
      type: string
      description: IRDAI unique identification number, e.g. 999N001V02 (product) or 999A007V01 (rider).
      pattern: '^[0-9]{3}[A-Z][0-9]{3}V[0-9]{2}$'
...
    ProductType:
      type: string
      enum: [TERM, TERM_ROP, NON_PAR_SAVINGS, PAR, ULIP]

    # --- Meta -------------------------------------------------------------------------------
...
    ProductStatus:
      type: string
      enum: [in_force, withdrawn]
    PptOption:
      type: string
      description: Premium-paying term option.
      enum: [regular, limited_10, single]
    Product:
      type: object
      required: [uin, name, category, status, entry_age_min, entry_age_max, maturity_age_max,
                 sa_min_inr, sa_max_inr, term_years_min, term_years_max, ppt_options,
                 benefit_payment_options, rider_uins, effective_from, effective_to, launch_enabled,
                 is_dummy, riders, documents]
      properties:
        uin: {$ref: '#/components/schemas/Uin'}
        name: {type: string}
        category: {$ref: '#/components/schemas/ProductType'}
        status: {$ref: '#/components/schemas/ProductStatus'}
        entry_age_min: {type: integer}
        entry_age_max: {type: integer}
        maturity_age_max: {type: integer}
        sa_min_inr: {$ref: '#/components/schemas/Money'}
        sa_max_inr:
          anyOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
        term_years_min: {type: integer, description: Lower bound of catalog.product.term_years.}
        term_years_max: {type: integer, description: Upper bound of catalog.product.term_years.}
        ppt_options:
          type: array
          items: {$ref: '#/components/schemas/PptOption'}
        benefit_payment_options:
          type: array
          description: >-
            How the benefit is paid, e.g. lumpsum or monthly_income (catalog.product.payout_options,
            renamed so the contract carries no payout field).
          items: {type: string}
        rider_uins:
          type: array
          items: {$ref: '#/components/schemas/Uin'}
        effective_from: {type: string, format: date}
        effective_to: {type: [string, 'null'], format: date}
        launch_enabled: {type: boolean}
        is_dummy: {type: boolean}
        riders:
          type: array
          items: {$ref: '#/components/schemas/Rider'}
        documents:
          type: array
          items: {$ref: '#/components/schemas/ProductDocument'}
    Rider:
      type: object
      required: [uin, name, attaches_to, sa_max_inr, is_dummy]
      properties:
        uin: {$ref: '#/components/schemas/Uin'}
        name: {type: string}
        attaches_to:
          type: array
          description: Base product UINs.
          items: {$ref: '#/components/schemas/Uin'}
        sa_max_inr:
          anyOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
        is_dummy: {type: boolean}
    ProductDocument:
      type: object
      required: [kind, version, language, uri, sha256]
      properties:
        kind:
          type: string
          enum: [CIS, BI, POLICY_WORDING]
        version: {type: string}
        language: {type: string}
        uri: {type: string}
        sha256: {$ref: '#/components/schemas/Sha256Hex'}
    KillSwitch:
      type: object
      required: [reason, actor]
      properties:
        reason: {type: string}
        actor: {type: string}
    KillSwitchResult:
      type: object
      required: [uin, status]
      properties:
        uin: {$ref: '#/components/schemas/Uin'}
        status: {$ref: '#/components/schemas/ProductStatus'}

    # --- Disclosure (TDD §3.8, §7.2) --------------------------------------------------------
...